% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.spline.R
\name{mcmc.spline}
\alias{mcmc.spline}
\title{MCMC sampling for location-scale bspline regression}
\usage{
mcmc.spline(
  m,
  it,
  burning,
  thinning,
  stepsize = NA,
  betastart = NA,
  gammastart = NA,
  epsstart = NA,
  taustart = NA
)
}
\arguments{
\item{m}{list with entries x and z, the predictors for mean and scale 
respectively and y, the response.}

\item{it}{integer   Number of iterations}

\item{burning}{integer   Number  of iterations to be burned}

\item{thinning}{integer   Number of iterations to be ignored between two
consecutive iterations}

\item{stepsize}{positive float stepsize for gamma sampler. Standard value is
sqrt(3) * (df)^(-1/6).}
}
\value{
MCMC sample for epsilon, tau, gamma and beta
}
\description{
MCMC sampling for location-scale bspline regression
}
