% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.spline}
\alias{plot.spline}
\title{plot method for spline objects}
\usage{
\method{plot}{spline}(m, sd = 1.96)
}
\arguments{
\item{m}{}

\item{sd}{}
}
\value{

}
\description{
plot method for spline objects
}
