% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.mcmcspline}
\alias{predict.mcmcspline}
\title{predict method for spline objects}
\usage{
\method{predict}{mcmcspline}(m, sample)
}
\arguments{
\item{m}{}

\item{sample}{}
}
\value{

}
\description{
predict method for spline objects
}
