% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.spline}
\alias{predict.spline}
\title{predict method for spline objects}
\usage{
\method{predict}{spline}(m, X, Z)
}
\arguments{
\item{m}{}

\item{X}{}

\item{Z}{}
}
\value{

}
\description{
predict method for spline objects
}
