% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mcmcspline}
\alias{print.mcmcspline}
\title{print method for mcmcspline objects}
\usage{
\method{print}{mcmcspline}(m, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{m}{}

\item{digits}{}

\item{...}{}
}
\value{

}
\description{
print method for mcmcspline objects
}
