% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.spline}
\alias{print.spline}
\title{print method for spline objects}
\usage{
\method{print}{spline}(m, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{m}{}

\item{digits}{}

\item{...}{}
}
\value{

}
\description{
print method for spline objects
}
