% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{spline}
\alias{spline}
\title{Location-Scale basis spline regression}
\usage{
spline(m, kn, order, p_order, smooth)
}
\arguments{
\item{m}{a list of the form list(x = x, z = x, y = y), with x the predictor
vector for the location, z, the predictor vector for the scale,
and y the response vector.}

\item{kn}{kn vector of integers 1. specifying number of knots for location
2. specifying number of knots for scale}

\item{order}{vector of integers 1. specifying order of spline for location
2. specifying order of spline for scale}

\item{p_order}{vector of integers 1. specifying order of spline for location
2. specifying order of spline for scale}

\item{smooth}{vector of positive floats specifying strength of penalization
1. for location parameters (1/tau)
2. for scale parameters (1/epsilon)}
}
\value{
a spline object with estimations for the location and scale
        coefficients.
}
\description{
The Location-scale bspline regression model enables flexible non-parametric
modeling of the mean and of the standard deviation. It is only possible to
model one predictor variable. To control the smoothness of the model a
penalization term is applied
}
