% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mcmcspline}
\alias{summary.mcmcspline}
\title{summary method for MCMC samples of a spline model}
\usage{
\method{summary}{mcmcspline}(model, par)
}
\arguments{
\item{par}{}

\item{m}{}
}
\value{

}
\description{
summary method for MCMC samples of a spline model
}
