% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.spline}
\alias{summary.spline}
\title{summary method for spline objects}
\usage{
\method{summary}{spline}(model, par)
}
\arguments{
\item{model}{}

\item{par}{}
}
\value{

}
\description{
summary method for spline objects
}
