% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.spline.R
\name{mcmc.spline}
\alias{mcmc.spline}
\title{MCMC sampling for location-scale bspline regression}
\usage{
mcmc.spline(m, it, burning, thinning, cov)
}
\arguments{
\item{m}{lmls spline object generated by a call to the spline_user_function
function}

\item{it}{integer   Number of iterations}

\item{burning}{integer   Number  of iterations to be burned}

\item{thinning}{integer   Number of iterations to be ignored between two
consecutive iterations}

\item{cov}{to be deleted}
}
\value{
MCMC sample for epsilon, tau, gamma and beta
}
\description{
MCMC sampling for location-scale bspline regression
}
