% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.mcmcspline}
\alias{plot.mcmcspline}
\title{plot function for MCMC samples of a spline model}
\usage{
\method{plot}{mcmcspline}(m, sample)
}
\arguments{
\item{m}{}

\item{sample}{}

\item{sd}{}
}
\value{

}
\description{
plot function for MCMC samples of a spline model
}
