% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.mcmcspline}
\alias{predict.mcmcspline}
\title{predict method for spline objects}
\usage{
\method{predict}{mcmcspline}(sample, m)
}
\arguments{
\item{sample}{sample from the mcmc.spline function}

\item{m}{model from the spline function}
}
\value{

}
\description{
predict method for spline objects
}
