% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.spline}
\alias{predict.spline}
\title{predict method for spline objects}
\usage{
\method{predict}{spline}(m, X, Z, isDesignmatrix = F)
}
\arguments{
\item{m}{}

\item{X}{vector of new x values}

\item{Z}{vector of new z values}

\item{isDesignmatrix}{Indication whether X and Z are already spline matrices}
}
\value{

}
\description{
predict method for spline objects
}
