% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{spline}
\alias{spline}
\title{Location-Scale basis spline regression}
\usage{
spline(m, kn, order, p_order, smooth)
}
\arguments{
\item{m}{A LMLS object generated by a call to the lmls function. Only one
predictor variable can be used and a predictor variable for the
scale parameter needs to be defined. Also light = True needs to be
set.}

\item{kn}{kn vector of integers 1. specifying number of knots for location
2. specifying number of knots for scale}

\item{order}{vector of integers 1. specifying order of spline for location
2. specifying order of spline for scale}

\item{p_order}{vector of integers 1. specifying order of spline for location
2. specifying order of spline for scale}

\item{smooth}{vector of positive floats specifying strength of penalization
1. for location parameters (1/tau)
2. for scale parameters (1/epsilon)}
}
\value{
a LMLS object with estimations for the location and scale
        coefficients.
}
\description{
The Location-scale bspline regression model enables flexible non-parametric
modeling of the mean and of the standard deviation. It is only possible to
model one predictor variable. To control the smoothness of the model a
penalization term is applied
}
